+++ packages/react-native-aria/listbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @react-native-aria/listbox
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-native-aria/interactions@0.2.15
+
 ## 0.2.7
 
 ### Patch Changes