+++ packages/unstyled/image-viewer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gluestack-ui/image-viewer
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-native-aria/interactions@0.2.15
+  - @gluestack-ui/utils@0.1.15
+
 ## 0.0.15
 
 ### Patch Changes