+++ packages/unstyled/link/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gluestack-ui/link
 
+## 0.1.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-native-aria/interactions@0.2.15
+  - @gluestack-ui/utils@0.1.15
+
 ## 0.1.27
 
 ### Patch Changes